fix: sonar issues

+++ src/main/java/com/esanchez/devel/breakscounter/util/Constants.java
@@ -8,9 +8,9 @@ public class Constants {
 	
 	public static final String APP_TITLE = "Breaks Counter";
 	
-	public static final int APP_WIDTH = 600;
+	public static final double APP_WIDTH = 600;
 	
-	public static final int APP_HEIGHT = 450;
+	public static final double APP_HEIGHT = 450;
 	
 	public static final String FONT_VERDANA = "Verdana";
 }
